+++ apollo-biz/src/main/java/com/ctrip/framework/apollo/biz/service/ItemSetService.java
@@ -16,6 +16,7 @@
  */
 package com.ctrip.framework.apollo.biz.service;
 
+import com.ctrip.framework.apollo.biz.config.BizConfig;
 import com.ctrip.framework.apollo.biz.entity.Audit;
 import com.ctrip.framework.apollo.biz.entity.Item;
 import com.ctrip.framework.apollo.biz.entity.Namespace;
@@ -25,6 +26,7 @@ import com.ctrip.framework.apollo.common.dto.ItemDTO;
 import com.ctrip.framework.apollo.common.exception.BadRequestException;
 import com.ctrip.framework.apollo.common.exception.NotFoundException;
 import com.ctrip.framework.apollo.common.utils.BeanUtils;
+import com.ctrip.framework.apollo.core.utils.StringUtils;
 import java.util.List;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -38,20 +40,23 @@ public class ItemSetService {
   private final CommitService commitService;
   private final ItemService itemService;
   private final NamespaceService namespaceService;
+  private final BizConfig bizConfig;
 
   public ItemSetService(
       final AuditService auditService,
       final CommitService commitService,
       final ItemService itemService,
-      final NamespaceService namespaceService) {
+      final NamespaceService namespaceService,
+      final BizConfig bizConfig) {
     this.auditService = auditService;
     this.commitService = commitService;
     this.itemService = itemService;
     this.namespaceService = namespaceService;
+    this.bizConfig = bizConfig;
   }
 
   @Transactional
-  public ItemChangeSets updateSet(Namespace namespace, ItemChangeSets changeSets){
+  public ItemChangeSets updateSet(Namespace namespace, ItemChangeSets changeSets) {
     return updateSet(namespace.getAppId(), namespace.getClusterName(), namespace.getNamespaceName(), changeSets);
   }
 
@@ -62,6 +67,16 @@ public class ItemSetService {
 
     if (namespace == null) {
       throw NotFoundException.namespaceNotFound(appId, clusterName, namespaceName);
+    }
+
+    if (bizConfig.isItemNumLimitEnabled()) {
+      int itemCount = itemService.findNonEmptyItemCount(namespace.getId());
+      int createItemCount = (int) changeSet.getCreateItems().stream().filter(item -> !StringUtils.isEmpty(item.getKey())).count();
+      int deleteItemCount = (int) changeSet.getDeleteItems().stream().filter(item -> !StringUtils.isEmpty(item.getKey())).count();
+      itemCount = itemCount + createItemCount - deleteItemCount;
+      if (itemCount > bizConfig.itemNumLimit()) {
+        throw new BadRequestException("The maximum number of items (" + bizConfig.itemNumLimit() + ") for this namespace has been reached. Current item count is " + itemCount + ".");
+      }
     }
 
     String operator = changeSet.getDataChangeLastModifiedBy();
@@ -84,7 +99,7 @@ public class ItemSetService {
 
     if (configChangeContentBuilder.hasContent()) {
       commitService.createCommit(appId, clusterName, namespaceName, configChangeContentBuilder.build(),
-                   changeSet.getDataChangeLastModifiedBy());
+                                 changeSet.getDataChangeLastModifiedBy());
     }
 
     return changeSet;